+++ include/mantella_bits/optimisationProblem/blackBoxOptimisationBenchmark.hpp
@@ -35,10 +35,10 @@ namespace mant {
 #if defined(MANTELLA_USE_PARALLEL_ALGORITHMS)
         friend class OptimisationAlgorithm;
             
-        std::vector<long double> serialise() const noexcept;
+        std::vector<double> serialise() const noexcept;
 
         void deserialise(
-            const std::vector<long double>& serialisedOptimisationProblem);
+            const std::vector<double>& serialisedOptimisationProblem);
 #endif
     };
 
@@ -154,13 +154,13 @@ namespace mant {
 
 #if defined(MANTELLA_USE_PARALLEL_ALGORITHMS)
     template <typename T>
-    std::vector<long double> BlackBoxOptimisationBenchmark<T>::serialise() const noexcept {
+    std::vector<double> BlackBoxOptimisationBenchmark<T>::serialise() const noexcept {
       return OptimisationProblem<T, T>::serialise();
     }
 
     template <typename T>
     void BlackBoxOptimisationBenchmark<T>::deserialise(
-        const std::vector<long double>& serialisedOptimisationProblem) {
+        const std::vector<double>& serialisedOptimisationProblem) {
       OptimisationProblem<T, T>::deserialise(serialisedOptimisationProblem);
     }
 #endif
