+++ include/mantella_bits/optimisationProblem/blackBoxOptimisationBenchmark/bentCigarFunction.hpp
@@ -22,10 +22,10 @@ namespace mant {
 #if defined(MANTELLA_USE_PARALLEL_ALGORITHMS)
         friend class OptimisationAlgorithm;
         
-        std::vector<long double> serialise() const noexcept;
+        std::vector<double> serialise() const noexcept;
 
         void deserialise(
-            const std::vector<long double>& serialisedOptimisationProblem);
+            const std::vector<double>& serialisedOptimisationProblem);
 #endif
     };
 
@@ -64,8 +64,8 @@ namespace mant {
     
 #if defined(MANTELLA_USE_PARALLEL_ALGORITHMS)
     template <typename T>
-    std::vector<long double> BentCigarFunction<T>::serialise() const noexcept {
-      std::vector<long double> serialisedOptimisationProblem = BlackBoxOptimisationBenchmark<T, T>::serialise();
+    std::vector<double> BentCigarFunction<T>::serialise() const noexcept {
+      std::vector<double> serialisedOptimisationProblem = BlackBoxOptimisationBenchmark<T, T>::serialise();
       
       for(std::size_t n = 0; n < rotationQ_.n_elem; ++n) {
         serialisedOptimisationProblem.push_back(rotationQ_(n));
@@ -76,7 +76,7 @@ namespace mant {
 
     template <typename T>
     void BentCigarFunction<T>::deserialise(
-        const std::vector<long double>& serialisedOptimisationProblem) {
+        const std::vector<double>& serialisedOptimisationProblem) {
       rotationQ_.set_size(this->numberOfDimensions_, this->numberOfDimensions_);
       for(std::size_t n = 0; n < rotationQ_.n_elem; ++n) {
         rotationQ_(n) = serialisedOptimisationProblem.pop_back();
